requirements.txt
Pillow
legacy-cgi
